+++ HISTORY.md
@@ -1,5 +1,12 @@
 # Version  history for psiutils
 
+Version 1.0.0 - 09 June 2026
+
+1. Change button button_text to text
+2. Add class IconButtonConfig
+3. Paste icon
+------------------------------
+
 Version 0.2.34 - 06 May 2026
 
 1. New icons
@@ -407,6 +414,7 @@ Version 0.2.23 - 03 February 2026
 ## 0.1.0 - 7 October 2024
 
 1. Initial version
+